requirements.txt
aiocache==0.11.1
aiocqhttp==1.3.0
aiofiles==0.5.0
aiohttp==3.6.2
aiorequests==0.1.1
aiosqlite==0.16.0
APScheduler==3.6.3
arrow==0.15.6
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.2.0
beautifulsoup4==4.9.0
binaryornot==0.4.4
blinker==1.4
Brotli==1.0.7
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cookiecutter==1.7.2
cryptography==3.3.1
demjson==2.2.4
distro==1.5.0
docker==4.4.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
expiringdict==1.2.0
fastapi==0.63.0
feedparser==5.2.1
Flask==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.6.30
httpcore==0.12.2
httptools==0.1.1
httpx==0.16.1
Hypercorn==0.10.1
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jinja2-time==0.2.0
jsonschema==3.2.0
loguru==0.5.3
MarkupSafe==1.1.1
multidict==4.7.5
nb-cli==0.3.1
nonebot-plugin-apscheduler==0.1.0
nonebot2==2.0.0a8.post1
numpy==1.19.4
opencc-python-reimplemented==0.1.5
paramiko==2.7.2
peewee==3.13.3
Pillow==8.2.0
poyo==0.5.0
priority==1.3.0
prompt-toolkit==1.0.14
pycparser==2.20
pydantic==1.7.3
pyfiglet==0.8.post1
Pygments==2.7.3
pygtrie==2.4.1
PyInquirer==1.0.3
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
Quart==0.11.5
regex==2020.11.13
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
soupsieve==2.0
starlette==0.13.6
text-unidecode==1.3
texttable==1.6.3
toml==0.10.1
typing==3.7.4.1
typing-extensions==3.7.4.2
tzlocal==2.0.0
urllib3==1.25.9
uvicorn==0.11.8
uvloop==0.14.0
wcwidth==0.2.5
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wsproto==0.15.0
yarl==1.4.2
psutil
cos-python-sdk-v5
qcloud-python-sts
jieba==0.42.1
wordcloud==1.8.1
pydub==0.25.1